+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 - **Example** added a json serialization and deserialization example of BarSeries using google-gson library
 - :tada: **Enhancement** added **`SqnCriterion`**
 - :tada: **Enhancement** added **`StandardDeviationCriterion`**
+- :tada: **Enhancement** added **`RelativeStandardDeviationCriterion`**
 - :tada: **Enhancement** added **`StandardErrorCriterion`**
 - :tada: **Enhancement** added **`VarianceCriterion`**